+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 
 ## [Unreleased]
+
 -   [DOC] Fix references and broken links in AUTHORS.rst. @loganthomas
 -   [DOC] Updated Broken links in the README and contributing docs. @nvamsikrishna05
 -   [INF] Update pre-commit hooks and remove mutable references. Issue #844. @loganthomas
@@ -14,7 +15,11 @@
 -   [DOC] Delete Read the Docs project and remove all readthedocs.io references from the repo. Issue #863. @loganthomas
 -   [DOC] Updated various documentation sources to reflect pyjanitor-dev ownership. @loganthomas
 -   [INF] Fix `isort` automatic checks. Issue #845. @loganthomas
+-   [ENH] `complete` function now uses variable args (*args) - @samukweku
+-   [EHN] Set `expand_column`'s `sep` default is `"|"`, same to `pandas.Series.str.get_dummies`. Issue #876. @Zeroto521
 -   [ENH] Deprecate `limit` from fill_direction. fill_direction now uses kwargs. @samukweku
+-   [ENH] Added `conditional_join` function that supports joins on non-equi operators. @samukweku
+-   [INF] Speed up pytest via `-n` (pytest-xdist) option. Issue #881. @Zeroto521
 -   [ENH] Fix warning message in `coalesce`, from bfill/fill;`coalesce` now uses variable arguments. @samukweku Issue  #882
 
 ## [v0.21.0] - 2021-07-16
@@ -225,7 +230,6 @@ who have helped make `pyjanitor`
 the package that it is today.
 
 [Unreleased]: https://github.com/pyjanitor-devs/pyjanitor/compare/v0.21.0...HEAD
-
 [v0.21.0]: https://github.com/pyjanitor-devs/pyjanitor/compare/v0.20.14...v0.21.0
 
 [v0.20.14]: https://github.com/pyjanitor-devs/pyjanitor/compare/v0.20.13...v0.20.14